thm2isis now uses out attrs and converted to func (#4504)
* thm2isis now uses out attrs and converted to func

* removed commented code that manually set pixel type

* added gtests

* better params

* IR test added

* reordered things

* cropped images

* cropped image for outattr tests too
